requirements.txt
Authlib==1.5.2
blinker==1.9.0
boto3==1.37.37
botocore==1.37.37
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
Flask==3.1.0
Flask-Login==0.6.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
s3transfer==0.11.5
six==1.17.0
urllib3==2.4.0
Werkzeug==3.1.3
